AI-RAG-BOT/requirements.txt
langchain==0.3.3
langchain-core==0.3.3
langchain-community==0.3.2
langchain-anthropic==0.3.0
langchain-huggingface==0.1.2          # ← important: new home for embeddings
faiss-cpu
pymupdf==1.24.10
sentence-transformers==3.1.1
streamlit==1.39.0
python-dotenv==1.0.1
tiktoken==0.8.0